cypress: removed repeated tests and enabled notebookbar tests
