+++ cypress_test/integration_tests/desktop/calc/top_toolbar_spec.js
@@ -1,4 +1,4 @@
-/* global describe it Cypress cy beforeEach require afterEach expect*/
+/* global describe it Cypress cy beforeEach require afterEach */
 
 var helper = require('../../common/helper');
 var desktopHelper = require('../../common/desktop_helper');
@@ -114,6 +114,7 @@ describe('Top toolbar tests.', function() {
 				});
 		});
 	});
+
 	it('Merge cells', function() {
 
 		// Select the full column
@@ -126,220 +127,6 @@ describe('Top toolbar tests.', function() {
 			.click();
 
 		desktopHelper.checkDialogAndClose('Merge Cells');
-	});
-
-
-	it('Remove cell border', function() {
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		//Add left border
-		cy.get('.w2ui-tb-image.w2ui-icon.frame02')
-			 .click({force: true});
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should('have.attr', 'style', 'border-left: 1px solid #000000');
-
-		// Then remove it
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame01')
-			 .click({force: true});
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should('not.have.attr', 'style');
-	});
-
-	it('Apply left border', function() {
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame02')
-			 .click({force: true});
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should('have.attr', 'style', 'border-left: 1px solid #000000');
-	});
-
-	it('Apply right border', function() {
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame03')
-			 .click({force: true});
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should('have.attr', 'style', 'border-right: 1px solid #000000');
-	});
-
-	it('Apply left and right border', function() {
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame04')
-			 .click();
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should('have.attr', 'style', 'border-left: 1px solid #000000; border-right: 1px solid #000000');
-	});
-
-	it('Apply top border', function() {
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame05')
-			 .click();
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should('have.attr', 'style', 'border-top: 1px solid #000000');
-	});
-
-	it('Apply bottom border', function() {
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame06')
-			 .click();
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should('have.attr', 'style', 'border-bottom: 1px solid #000000');
-	});
-
-	it('Apply top and bottom border', function() {
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame07')
-			 .click();
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should('have.attr', 'style', 'border-top: 1px solid #000000; border-bottom: 1px solid #000000');
-	});
-
-	it('Apply border for all sides', function() {
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame08')
-			 .click();
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should('have.attr', 'style', 'border-top: 1px solid #000000; border-bottom: 1px solid #000000; border-left: 1px solid #000000; border-right: 1px solid #000000');
-	});
-
-	it('Apply horizontal borders for multiple cells', function() {
-		calcHelper.selectEntireSheet();
-
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame09')
-			 .click();
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should(function(cells) {
-				expect(cells).to.have.lengthOf(4);
-				for (var i = 0; i < cells.length; i++) {
-					expect(cells[i]).to.have.attr('style', 'border-top: 1px solid #000000; border-bottom: 1px solid #000000');
-				}
-			});
-	});
-
-	it('Apply horizontal inner borders and vertical outer borders', function() {
-		calcHelper.selectEntireSheet();
-
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame10')
-			 .click();
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should(function(cells) {
-				expect(cells).to.have.lengthOf(4);
-				for (var i = 0; i < cells.length; i++) {
-					if (i == 0 || i == 2)
-						expect(cells[i]).to.have.attr('style', 'border-top: 1px solid #000000; border-bottom: 1px solid #000000; border-left: 1px solid #000000');
-					else
-						expect(cells[i]).to.have.attr('style', 'border-top: 1px solid #000000; border-bottom: 1px solid #000000');
-				}
-			});
-	});
-
-	it('Apply vertical inner borders and horizontal outer borders', function() {
-		calcHelper.selectEntireSheet();
-
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame11')
-			 .click();
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should(function(cells) {
-				expect(cells).to.have.lengthOf(4);
-				for (var i = 0; i < cells.length; i++) {
-					if (i == 0 || i == 1)
-						expect(cells[i]).to.have.attr('style', 'border-top: 1px solid #000000; border-left: 1px solid #000000; border-right: 1px solid #000000');
-					else
-						expect(cells[i]).to.have.attr('style', 'border-left: 1px solid #000000; border-right: 1px solid #000000');
-				}
-			});
-	});
-
-	it('Apply all inner and outer borders', function() {
-		calcHelper.selectEntireSheet();
-
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame12')
-			 .click();
-
-		calcHelper.selectEntireSheet();
-
-		cy.get('#copy-paste-container table td')
-			.should(function(cells) {
-				expect(cells).to.have.lengthOf(4);
-				for (var i = 0; i < cells.length; i++) {
-					expect(cells[i]).to.have.attr('style', 'border-top: 1px solid #000000; border-bottom: 1px solid #000000; border-left: 1px solid #000000; border-right: 1px solid #000000');
-				}
-			});
-	});
-
-	it('Apply More Border styles', function() {
-		cy.get('#tb_editbar_item_setborderstyle')
-			.click();
-
-		cy.get('.w2ui-tb-image.w2ui-icon.frame13')
-		    .click();
-
-		 desktopHelper.checkDialogAndClose('Format Cells');
 	});
 
 	it('Clear Direct formatting.', function() {
