+++ cypress_test/integration_tests/desktop/impress/top_toolbar_spec.js
@@ -138,4 +138,44 @@ describe('Top toolbar tests.', function() {
 
 		cy.get('.leaflet-pane.leaflet-overlay-pane g.Page .TextParagraph .TextPosition').should('have.attr', 'x', '1400');
 	});
+
+	it('Apply superscript on selected text.', function() {
+		impressHelper.selectTextOfShape();
+
+		cy.get('.leaflet-pane.leaflet-overlay-pane g.Page .TextPosition')
+			.should('have.attr', 'y', '8643');
+
+		cy.get('.leaflet-pane.leaflet-overlay-pane g.Page .TextParagraph')
+			.should('have.attr', 'font-size', '1129px');
+
+		helper.typeIntoDocument('{ctrl}{shift}p');
+
+		impressHelper.triggerNewSVGForShapeInTheCenter();
+
+		cy.get('.leaflet-pane.leaflet-overlay-pane g.Page .TextPosition')
+			.should('have.attr', 'y', '8271');
+
+		cy.get('.leaflet-pane.leaflet-overlay-pane g.Page .TextParagraph')
+			.should('have.attr', 'font-size', '655px');
+	});
+
+	it('Apply subscript on selected text.', function() {
+		impressHelper.selectTextOfShape();
+
+		cy.get('.leaflet-pane.leaflet-overlay-pane g.Page .TextPosition')
+			.should('have.attr', 'y', '8643');
+
+		cy.get('.leaflet-pane.leaflet-overlay-pane g.Page .TextParagraph')
+			.should('have.attr', 'font-size', '1129px');
+
+		helper.typeIntoDocument('{ctrl}{shift}b');
+
+		impressHelper.triggerNewSVGForShapeInTheCenter();
+
+		cy.get('.leaflet-pane.leaflet-overlay-pane g.Page .TextPosition')
+			.should('have.attr', 'y', '8734');
+
+		cy.get('.leaflet-pane.leaflet-overlay-pane g.Page .TextParagraph')
+			.should('have.attr', 'font-size', '655px');
+	});
 });
