+++ cypress_test/plugins/blacklists.js
@@ -202,7 +202,22 @@ var updateScreenshotList = [
 	['desktop/calc/help_dialog_update_spec.js', []],
 ];
 
+var notebookbarOnlyList = [
+	'desktop/writer/top_toolbar_spec.js',
+	'desktop/calc/cell_appearance_spec.js',
+	'desktop/writer/table_operation_spec.js',
+	'desktop/impress/table_operation_spec.js',
+	'desktop/writer/annotation_spec.js',
+	'desktop/calc/annotation_spec.js',
+	'desktop/impress/annotation_spec.js',
+	'desktop/calc/row_column_operation_spec.js',
+	'desktop/impress/image_operation_spec.js',
+	'desktop/calc/image_operation_spec.js',
+	'desktop/writer/image_operation_spec.js',
+];
+
 module.exports.nextcloudBlackList = nextcloudBlackList;
 module.exports.nextcloudOnlyList = nextcloudOnlyList;
 module.exports.phpProxyBlackList = phpProxyBlackList;
 module.exports.updateScreenshotList = updateScreenshotList;
+module.exports.notebookbarOnlyList = notebookbarOnlyList;
