--- cypress_test/plugins/whitelists.js
@@ -1,15 +1,0 @@
-var notebookbarOnlyList = [
-	// 'desktop/writer/top_toolbar_spec.js',
-	// 'desktop/calc/cell_appearance_spec.js',
-	// 'desktop/writer/table_operation_spec.js',
-	// 'desktop/impress/table_operation_spec.js',
-	'desktop/writer/annotation_spec.js',
-	'desktop/calc/annotation_spec.js',
-	'desktop/impress/annotation_spec.js',
-	'desktop/calc/row_column_operation_spec.js',
-	'desktop/impress/image_operation_spec.js',
-	'desktop/calc/image_operation_spec.js',
-	'desktop/writer/image_operation_spec.js',
-];
-
-module.exports.notebookbarOnlyList = notebookbarOnlyList;
